repository tests

+++ src/main/java/be/vdab/justgetit/repositories/ProductRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
 public interface ProductRepository extends JpaRepository <Product, Long> {
     List<Product> findByVerkoopprijsBetween(BigDecimal van , BigDecimal tot);
     List<Product> findByBesteldIsNull();
-    List<Product> findByVoorraadLessThan(@Param("aantal") BigDecimal aantal);
+    List<Product> findByVoorraadLessThan(@Param("aantal") int aantal);
     List<Product>findByNaamContaining(String zoekString);
 }
